+++ app/src/js/reducers/granules.js
@@ -150,9 +150,12 @@ export default createReducer(initialState, {
   [GRANULE_DELETE_CLEAR_ERROR]: createClearItemReducer('deleted'),
 
   [SEARCH_GRANULES]: (state, action) => {
-    state.list.params.infix = action.infix;
+    state.list.params.prefix = action.infixBoolean ? null : action.xfix;
+    state.list.params.infix = action.infixBoolean ? action.xfix : null;
   },
   [CLEAR_GRANULES_SEARCH]: (state) => {
+    // in general, we want to clear both prefix and infix
+    state.list.params.prefix = null;
     state.list.params.infix = null;
   },
   [FILTER_GRANULES]: (state, action) => {