+++ cypress/integration/granules_spec.js
@@ -301,12 +301,44 @@ describe('Dashboard Granules Page', () => {
     });
 
     it('Should update URL and table when search filter is changed.', () => {
-      const infix = 'A0142558';
+      const prefix = 'test_';
       cy.visit('/granules');
       cy.get('.search').as('search');
-      cy.get('@search').click().type(infix);
-      cy.url().should('include', `search=${infix}`);
+      cy.get('@search').click().type(prefix);
+      cy.url().should('include', `search=${prefix}`);
       cy.get('.table .tbody .tr').should('have.length', 1);
+      cy.get('.table .tbody .tr').eq(0).children('.td').eq(2)
+        .contains(prefix);
+    });
+
+    it('Should update results when toggling search by infix', () => {
+      const prefix = 'MOD09GQ.A241';
+      const infix = 'A153';
+      cy.visit('/granules');
+      cy.get('#chk_isInfixSearch').should('not.be.checked');
+      cy.get('.search').as('search');
+      cy.get('@search').click().type(prefix).type('{enter}');
+      cy.get('.table .tbody .tr').should('have.length', 1);
+      cy.get('@search').click().type(infix).type('{enter}');
+      cy.get('.table .tbody .tr').should('have.length', 0);
+      cy.get('#chk_isInfixSearch').click({ force: true }).should('be.checked');
+      cy.get('.table .tbody .tr').should('have.length.at.least', 1);
+      cy.get('.table .tbody .tr').eq(0).children('.td').eq(2)
+        .contains(infix);
+    });
+
+    it('Should have search by infix input in granule status list view and update results when toggling search by infix', () => {
+      const prefix = 'test_';
+      const infix = 'A201';
+      cy.visit('/granules/completed');
+      cy.get('#chk_isInfixSearch').should('not.be.checked');
+      cy.get('#search').as('search');
+      cy.get('@search').click().type(prefix).type('{enter}');
+      cy.get('.table .tbody .tr').should('have.length', 1);
+      cy.get('@search').click().type(infix).type('{enter}');
+      cy.get('.table .tbody .tr').should('have.length', 0);
+      cy.get('#chk_isInfixSearch').click({ force: true }).should('be.checked');
+      cy.get('.table .tbody .tr').should('have.length.at.least', 1);
       cy.get('.table .tbody .tr').eq(0).children('.td').eq(2)
         .contains(infix);
     });