+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **security vulnerabilities**
   - Overrode pbkdf2 package and upgraded related packages to fix critical audit issue
 - **CUMULUS-4039**
+  - Updated to allow slashes in provider name and collection name
+- **CUMULUS-4122**
+  - Improved granule search performance by using prefix matching by default.
+  - Added a `Search by Infix` toggle to allow infix (contains) searches for Granule ID
   - Updated to allow slashes in provider name and collection name 
 - **CUMULUS-4171**
   - Removed override of pbkdf2 ^3.1.3 for crypto-browserify since pbkdf2 ^3.1.3 will be pulled automatically